Add dynamic page generation for each blog and link out to each on the blog page
